move download_location fixture to conftest
